Fixed the state persisting Moved classes into individual files Tidied up some code
