--- README.md
@@ -1,2 +1,0 @@
-# WallGame
-Project to create a connecring wall game
