+++ group.py
@@ -1,0 +1,8 @@
+class Group:
+
+    solved = "n"
+    solvedOrder = 0
+
+    def __init__(self,groupID,connexion):
+        self.groupID = groupID
+        self.connexion = connexion
