--- wall20200503sdp1TEST.txt
@@ -1,20 +1,0 @@
-Wookie
-Jawa
-Tuskan Raider
-Ewok
-Flamethrower
-Blow Gun 
-Rapier
-Stilleto
-Beatles
-Rolling Stones
-The Who
-Led Zeppelin
-Mississipi
-Colorado
-Missouri 
-St Lawrence
-Races in Star Wars
-Deadly Weapons
-Bands
-N.American Rivers
